Add MontMarte service tests and test script
